+++ apps/bam.iobio/src/index.d.ts
@@ -24,7 +24,6 @@ declare module 'iobio-demo-app';
 declare namespace JSX {
 	interface IntrinsicElements {
 		'iobio-percent-box': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
-		'iobio-histogram': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
 		'iobio-data-broker': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
 		'iobio-coverage-depth': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
 	}
