+++ components/src/index.d.ts
@@ -20,3 +20,12 @@
  */
 
 declare module 'components';
+
+declare namespace JSX {
+	interface IntrinsicElements {
+		'iobio-percent-box': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
+		'iobio-histogram': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
+		'iobio-data-broker': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
+		'iobio-coverage-depth': DetailedHTMLProps<HTMLAttributes<HTMLElement>, HTMLElement>;
+	}
+}
